+++ app/services/manual_solver/types.py
@@ -11,12 +11,12 @@ class OptimizationEvent(TypedDict, total=False):
     iteration: int
     elapsed_ms: float
     timestamp: datetime
-    current_distance_in_meters: int
-    current_duration_in_seconds: int
-    best_distance_in_meters: int
-    best_duration_in_seconds: int
-    distance_improvement_in_meters: int
-    duration_improvement_in_seconds: int
+    current_distance_in_meters: float
+    current_duration_in_seconds: float
+    best_distance_in_meters: float
+    best_duration_in_seconds: float
+    distance_improvement_in_meters: float
+    duration_improvement_in_seconds: float
     improvement_percent: float
     iterations_without_improvement: int
     objective_value: float
@@ -33,12 +33,12 @@ class OptimizationEvent(TypedDict, total=False):
 @dataclass
 class Assignment:
     tour: List[int]
-    total_distance_in_meters: int
-    total_duration_in_seconds: int
+    total_distance_in_meters: float
+    total_duration_in_seconds: float
     algorithm_used: str
     elapsed_ms: float
     iterations: int = 0
-    history: List[int] = field(default_factory=list[int])
+    history: List[float] = field(default_factory=list[float])
     logs: List[OptimizationEvent] = field(default_factory=list[OptimizationEvent])
 
 
@@ -84,8 +84,8 @@ class RoutingSearchParameters:
 
 @dataclass
 class ManualSolverProblem:
-    distance_matrix: List[List[int]]
-    time_matrix: List[List[int]]
+    distance_matrix: List[List[float]]
+    time_matrix: List[List[float]]
     depot: int = 0
     
 ComparisonScenario = Literal["no_improvement", "with_improvement"]
